+++ rse/client/search.py
@@ -11,6 +11,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 from rse.main import Encyclopedia
 from rse.logger import bot
 import logging
+import sys
 
 
 def main(args, extra):