+++ CHANGELOG.md
@@ -14,4 +14,5 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip.
 
 ## [0.2.x](https://github.com/rse/rse/tree/master) (0.0.x)
+ - first draft including database backends and GitHub parser (0.0.1)
  - skeleton release (0.0.0)